Rest of Projectile actor refactor
