+++ Assets/Code/Unity/Actor/CardEntities/HandProjectile/ProjectileDieOnCollision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6043f346effd92d489cb4f3a4afa6d79
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
