+++ Assets/Code/Unity/Tests/Actor/CardEntities/ProjectileDieOnCollisionTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 465e19dbcbd81b14a97ad9d33a25e8e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
