+++ Assets/Resources/Prefabs/HandInteractable.prefab
@@ -416,6 +416,7 @@ GameObject:
   - component: {fileID: 7803066593573095463}
   - component: {fileID: 2298003805780703171}
   - component: {fileID: 4134921874225974522}
+  - component: {fileID: 7029709375160896283}
   - component: {fileID: 6483888452169677458}
   - component: {fileID: 2800447118953497448}
   m_Layer: 12
@@ -1301,6 +1302,18 @@ MonoBehaviour:
     serializedVersion: 2
     m_Bits: 8192
   maxSpeed: 0.8
+--- !u!114 &7029709375160896283
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5390251873298114089}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6043f346effd92d489cb4f3a4afa6d79, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &6483888452169677458
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1316,8 +1329,8 @@ MonoBehaviour:
   OnCollsionEntered:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 5390251873298114091}
-        m_TargetAssemblyTypeName: SaloonSlingers.Unity.Actor.HandProjectile, SaloonSlingers.Unity
+      - m_Target: {fileID: 7029709375160896283}
+        m_TargetAssemblyTypeName: SaloonSlingers.Unity.Actor.DieOnCollision, SaloonSlingers.Unity
         m_MethodName: HandleCollision
         m_Mode: 0
         m_Arguments:
@@ -1334,8 +1347,8 @@ MonoBehaviour:
   OnTriggerEntered:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 5390251873298114091}
-        m_TargetAssemblyTypeName: SaloonSlingers.Unity.Actor.HandProjectile, SaloonSlingers.Unity
+      - m_Target: {fileID: 7029709375160896283}
+        m_TargetAssemblyTypeName: SaloonSlingers.Unity.Actor.DieOnCollision, SaloonSlingers.Unity
         m_MethodName: HandleCollision
         m_Mode: 0
         m_Arguments:
